requirements.txt
streamlit
pandas
google-cloud-bigquery
langchain-google-vertexai
langchain-core
db-dtypes
tabulate
